llm/openai/requirements.txt
ipykernel
openai
python-dotenv
yfinance
pandas
matplotlib
colorama
streamlit